fix/rename type file for components
