+++ src/widgets/Accordeon/index.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react';
 import classNames from 'classnames';
 
 import cross from '../../assets/images/added_cross.svg';
-import { AccordionProps } from './types';
+import { AccordionProps } from './type';
 import styles from './styles.module.scss';
 
 const Accordion: FC<AccordionProps> = ({
