+++ src/widgets/Accordeon/type.ts
